+++ src/OpenLoco/src/Paint/PaintTrack.cpp
@@ -41,7 +41,7 @@ namespace OpenLoco::Paint
         }
         const auto height = elTrack.baseZ() * 4;
         const auto rotation = (session.getRotation() + elTrack.unkDirection()) & 0x3;
-        if (((session.getViewFlags() & Ui::ViewportFlags::height_marks_on_land) != Ui::ViewportFlags::none)
+        if (((session.getViewFlags() & Ui::ViewportFlags::height_marks_on_tracks_roads) != Ui::ViewportFlags::none)
             && session.getRenderTarget()->zoomLevel == 0)
         {
             const bool isLast = elTrack.isFlag6();